make shebang line more cross platform by using env bash, make istio version check less tedious to maintain, drop repo check
Macs tend to have bash be in differing spots, especially some devs who
want to substitute in a newer version. Use /usr/bin/env bash to
basically use the bash defined by the environment.

Don't make it so we have to define specific istio versions to support.
We know we *should* support anything newer than 1.22.1 and 1.23.0, so
block around that. We shouldn't break in newer versions.

Drop the repo check. The script should be able to handle any hub used
especially so it can be picked up and used by the Istio build pipeline
which will need to set this to its private image repos.

Signed-off-by: Daniel Hawton <daniel@hawton.org>